\name{CpG.pos}
\alias{CpG.pos}
\docType{data}
\title{
Position of CpG Sites for the methylation Data
}
\description{
This dataset contains pseudo positions of 200 CpG sites.
}
\usage{data(CpG.pos)}
\format{
  The format is:
  CpG sites by rows; column 1: names of CpG sites; column 2: positions of CpG sites.
}

\keyword{datasets}
