\name{CpG_pos}
\alias{CpG_pos}
\docType{data}
\title{
Pseudo position of CpG sites for the example data.
}
\description{
It contains positions of 200 CpG sites.
}
\usage{data(CpG_pos)}
\format{
  The format is:
  200 CpG sites by rows; Names and positions of CpG sites by columns.
}

\keyword{datasets}
