\name{diabetes.geno}
\alias{diabetes.geno}
\docType{data}
\title{
Genotype Data of Candidate Diabetes Genes
}
\description{
A data frame contains 23 SNPs for 115 individuals.
}
\usage{data(diabetes.geno)}
\format{
 The format is:
 subjects by rows and genotypes by columns.
}
\references{Wang, M. H., Li, J., Yeung, V. S. Y., Zee, B. C. Y., Yu, R. H. Y., Ho, S., & Waye, M. M. Y. (2014). Four pairs of gene-gene interactions associated with increased risk for type 2 diabetes (CDKN2BAS-KCNJ11), obesity (SLC2A9-IGF2BP2, FTO-APOA5), and hypertension (MC4R-IGF2BP2) in Chinese women. Meta Gene, 2, 384-391. http://doi.org/10.1016/j.mgene.2014.04.010}

\keyword{datasets}
