\name{genotype}
\alias{genotype}
\docType{data}
\title{
An example genotype dataset.
}
\description{
This simulated dataset contains 300 individuals and 200 SNPs.
}
\usage{data(genotype)}
\format{
  The format is:
  300 subjects by rows and its corresponding genotype variables by columns.
}

\keyword{datasets}
