% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf.R
\name{maf}
\alias{maf}
\title{Minor allele frequency}
\usage{
maf(data, which.marker = NULL)
}
\arguments{
\item{data}{a data frame or matrix contains genotypes in the columns. Genotypes should be coded as (0, 1, 2) or (0, 1).}

\item{which.marker}{a numeric vector with length = w.order indicates which SNP to calculate the MAF. When which.marker = NULL, MAF of all the markers is calculated. Default is NULL.}
}
\value{
The MAF of one marker
}
\description{
Function to calculate minor allele frequencies.
}
\examples{
data(mydata)
result<-maf(mydata, which.marker=10)
}
