% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation.recode.R
\name{methylation.recode}
\alias{methylation.recode}
\title{Recode methylation data}
\usage{
methylation.recode(data)
}
\arguments{
\item{data}{a data frame or matrix contains methylation data in the columns.}
}
\description{
{This function performs redocing for methylation data by a K-means clustering method. The methylation data set will be recoded into two (high and low) levels according to the methylated level of each CpG site.}
}
\examples{
data(methylation)
data.recoded <- methylation.recode(methylation)
}
