\name{methylation}
\alias{methylation}
\docType{data}
\title{
An example methylation dataset.
}
\description{
This simulated dataset contains 300 individuals and 200 CpG sites.
}
\usage{data(methylation)}
\format{
  The format is:
  300 subjects by rows and its corresponding methylation variables by columns.
}

\keyword{datasets}
