\name{phenotype2}
\alias{phenotype2}
\docType{data}
\title{
A simulated response variable of the genotype-methylation dataset.
}
\description{
A binary variable indicate the status of 300 individuals (1 = affacted, 0 = unaffacted}
\usage{data(phenotype2)}
\format{
  The format is:
 300 rows and 1 column.
}

\keyword{datasets}
