% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf.R
\name{maf}
\alias{maf}
\title{Minor Allele Frequency}
\usage{
maf(data, which.snp = NULL)
}
\arguments{
\item{data}{a data frame or matrix containing genotypes in the columns. Genotypes should be coded as (0, 1, 2) or (0, 1).}

\item{which.snp}{a numeric value, indicating which SNP to calculate. When which.snp = NULL, MAF of all the markers is calculated. Default is NULL.}
}
\value{
The MAF of one marker.
}
\description{
Calculate minor allele frequency.
}
\examples{
data(diabetes.geno)
result <- maf(diabetes.geno, which.snp=10)
}
