% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation.recode.R
\name{methylation.recode}
\alias{methylation.recode}
\title{Recode Methylation Data}
\usage{
methylation.recode(data)
}
\arguments{
\item{data}{a data frame or matrix contains methylation data in the columns.}
}
\description{
{Code a CpG variable into two levels (high and low) by the two-mean clustering method.}
}
\examples{
data(methylation)
data.recoded <- methylation.recode(methylation)
}
